implement #3 and split the code to several files
