+++ karpov/assets/meta-header-footer.html
@@ -1,0 +1,38 @@
+{{define "head"}}
+<meta charset="UTF-8">
+<meta name="viewport" content="width=device-width, initial-scale=1.0">
+<meta http-equiv="X-UA-Compatible" content="ie=edge">
+{{end}}
+
+{{define "header"}}
+<header>
+    <nav class="uk-navbar-container" uk-navbar>
+        <div class="uk-navbar-left">
+            <ul class="uk-navbar-nav">
+                <li class="uk-active"><a href="/posts">Статьи</a></li>
+                <li>
+                    <a href="#">Работа со статьями</a>
+                    <div class="uk-navbar-dropdown">
+                        <ul class="uk-nav uk-navbar-dropdown-nav">
+                            <li><a href="/posts/new">Новая <span uk-icon="plus"></span></a></li>
+                            <li><a href="#">Удалить</a></li>
+                            <li><a href="#">Редактировать?</a></li>
+                        </ul>
+                    </div>
+                </li>
+                <li><a href="#">И еще</a></li>
+            </ul>
+
+        </div>
+    </nav>
+</header>
+{{end}}
+
+{{define "footer"}}
+<!-- UIkit CSS -->
+<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/uikit/3.2.0/css/uikit.min.css" />
+
+<!-- UIkit JS -->
+<script src="https://cdnjs.cloudflare.com/ajax/libs/uikit/3.2.0/js/uikit.min.js"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/uikit/3.2.0/js/uikit-icons.min.js"></script>
+{{end}}
