+++ karpov/assets/postNew.html
@@ -1,0 +1,60 @@
+{{define "indexNewPOST"}}
+<!DOCTYPE html>
+<html lang="ru">
+
+<head>
+    {{template "head"}}
+    <title>{{.Title}}</title>
+</head>
+
+<body>
+    <div class="uk-container uk-width-5-6">
+        <!-- HEADER -->
+        {{template "header"}}
+        <!-- CONTENT -->
+        <div class="uk-text-center" uk-grid>
+            <div class="uk-width-1-5">
+                <div class="uk-card uk-card-default uk-card-body">Left</div>
+            </div>
+            <div class="uk-width-3-5">
+                <div class="uk-card uk-card-default uk-card-body">
+                    {{template "newPost" .Post}}
+                </div>
+            </div>
+            <div class="uk-width-1-5">
+                <div class="uk-card uk-card-default uk-card-body">Right</div>
+            </div>
+        </div>
+        <!-- FOOTER -->
+        {{template "footer"}}
+    </div>
+</body>
+
+</html>
+{{end}}
+
+{{define "newPost"}}
+<form action="/posts" method="POST" id="formPostEdit">
+    <fieldset class="uk-fieldset">
+
+        <legend class="uk-legend">Создание новой статьи</legend>
+
+        <div class="uk-margin">
+            <input class="uk-input" name="title" type="text" placeholder="Input" value="{{.Title}}">
+        </div>
+
+        <div class="uk-margin">
+            <select class="uk-select" name="rubric">
+                <option value="Go for fun">Go for fun</option>
+                <option value="Go for cry">Go for cry</option>
+            </select>
+        </div>
+
+        <div class="uk-margin">
+            <textarea class="uk-textarea" rows="10" placeholder="blog content" name="content">{{.Content}}</textarea>
+        </div>
+
+    </fieldset>
+    <button type="submit" form="formPostEdit" class="uk-button uk-button-default">Save</button>
+</form>
+{{end}}
